[FE-1172] breaking down in components
