+++ src/pages/domains/components/index.js
@@ -2,10 +2,19 @@ import Container from './Container';
 import CreateForm from './CreateForm';
 import SendingDomainsTab from './SendingDomainsTab';
 import TrackingDomainsTab from './TrackingDomainsTab';
-
+import SendingDomainsTable from './SendingDomainsTable';
+import TrackingDomainsTable from './TrackingDomainsTable';
+import SetupForSending from './SetupForSending';
+import DomainStatusSection from './DomainStatusSection';
+import DeleteDomainSection from './DeleteDomainSection';
 export default {
   CreateForm,
   Container,
   SendingDomainsTab,
   TrackingDomainsTab,
+  SendingDomainsTable,
+  TrackingDomainsTable,
+  SetupForSending,
+  DeleteDomainSection,
+  DomainStatusSection,
 };
